requirements.txt
aiopg
PyGithub
sanic
jinja2
aiohttp
pyyaml
uvloop
